Fixing CNAM querying logic to be more efficient.
This commit makes it so that we DON'T do both a free AND paid lookup, and
instead respects the user's settings and does the appropriate CNAM query.

This speeds up processing as well as keep things fast (for OpenCNAM and users).

I've also added a 5 second timeout--incase something awful happens, no calls
will suffer.